Use updated generator and modify status reconciliation logic
