+++ pkg/resource/record_set/manager_factory.go
@@ -82,7 +82,7 @@ func (f *resourceManagerFactory) IsAdoptable() bool {
 // RequeueOnSuccessSeconds returns true if the resource should be requeued after specified seconds
 // Default is false which means resource will not be requeued after success.
 func (f *resourceManagerFactory) RequeueOnSuccessSeconds() int {
-	return 60
+	return 0
 }
 
 func newResourceManagerFactory() *resourceManagerFactory {
