+++ templates/hooks/record_set/sdk_read_many_post_set_output.go.tpl
@@ -1,10 +1,11 @@
 
-    // Status represents whether record changes have been fully propagated to all
-    // Route 53 authoritative DNS servers. The current status for the propagation
-    // should be updated per reconciliation
-    if ko.Status.Status != nil {
-        err = rm.syncStatus(ctx, ko)
-        if err != nil {
-            return nil, err
-        }
-    }
+	// Status represents whether record changes have been fully propagated to all
+	// Route 53 authoritative DNS servers. The current status for the propagation
+	// should be updated if it's not already in an INSYNC state
+	err = rm.syncStatus(ctx, ko)
+	if err != nil {
+		return nil, err
+	}
+	if ko.Status.Status == nil || *ko.Status.Status == svcsdk.ChangeStatusPending {
+		ackcondition.SetSynced(&resource{ko}, corev1.ConditionFalse, nil, nil)
+	}
